Extend Decidim Initiatives AdminLog Presenter
